sumo update

+++ database/migrations/2022_05_23_040934_create_accountants_table.php
@@ -21,8 +21,8 @@ class CreateAccountantsTable extends Migration
 			$table->timestamp('dob');
 			$table->string('email');
 			$table->string('password');
-			$table->integer('dept_id')->default('1');
-			$table->integer('role_id');
+			$table->unsignedBigInteger('dept_id')->default('1');
+			$table->unsignedBigInteger('role_id');
 			$table->boolean('status');
 		});
 	}
